refactor(diagnostic): move diagnostic UX out of core nono crate
Relocate DiagnosticFormatter and all user-facing diagnostic rendering from
the core 'nono' crate to 'nono-cli', restoring the library/CLI boundary
documented in AGENTS.md: the library stays a pure sandbox primitive and the
CLI owns policy, footer text, flag suggestions, and stderr heuristics.

Core 'nono::diagnostic' now exposes only structured, policy-free facts:
DenialReason, DenialRecord, IpcDenialRecord, SandboxViolation, and the
seatbelt_operation_to_access mapping. Everything else (DiagnosticFormatter,
PolicyExplanation, ErrorObservation, analyze_error_output, and the footer
rendering with 'nono why'/'nono learn' guidance and CLI flag suggestions)
moves to crates/nono-cli/src/diagnostic/.

The bulk of this change is a near-verbatim move; the only behavioral edits:

- Remove dead code surfaced by the move (previously masked by pub visibility
  in the library): PolicyExplanation.details/policy_source fields (written but
  never read), and the unused DiagnosticFormatter methods
  with_blocked_protected_file, detect_protected_file_in_error, and
  format_summary.
- Deduplicate suggested_flag_parts: the formatter now uses the single
  query_ext implementation, which guards against suggesting access to the
  root filesystem. This fixes an edge case where a nonexistent path directly
  under '/' could yield a '--allow /' suggestion.

No change to diagnostic footer output in normal cases; the moved unit tests
are the rendering contract and pass unchanged.
