+++ crates/nono/src/lib.rs
@@ -67,10 +67,7 @@ pub use capability::{
     AccessMode, CapabilitySet, CapabilitySource, FsCapability, IpcMode, NetworkMode,
     ProcessInfoMode, SignalMode, SocketScope, UnixSocketCapability, UnixSocketMode, UnixSocketOp,
 };
-pub use diagnostic::{
-    CommandContext, DenialReason, DenialRecord, DiagnosticFormatter, DiagnosticMode,
-    IpcDenialRecord, SandboxViolation,
-};
+pub use diagnostic::{DenialReason, DenialRecord, IpcDenialRecord, SandboxViolation};
 pub use error::{NonoError, Result};
 pub use keystore::{
     LoadedSecret, is_apple_password_uri, is_bw_uri, is_env_uri, is_file_uri, is_keyring_uri,
